feat: put in a component
